+++ spec/eve_online/esi/models/dogma_attributes_spec.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+require 'spec_helper'
+
+describe EveOnline::ESI::Models::DogmaAttributes do
+  it { should be_a(EveOnline::ESI::Models::Base) }
+
+  let(:options) { double }
+
+  subject { described_class.new(options) }
+
+  describe '#initialize' do
+    its(:options) { should eq(options) }
+  end
+
+  describe '#dogma_attributes' do
+    let(:dogma_attribute_short) { double }
+
+    let(:option) { double }
+
+    let(:options) { [option] }
+
+    before { expect(DogmaAttributeShort).to receive(:new).with(option).and_return(dogma_attribute_short) }
+
+    before { expect(subject.dogma_attributes).to eq([dogma_attribute_short]) }
+  end
+end